requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==0.7
cron-descriptor==1.4.0
cryptography==41.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.2
django-appconf==1.0.5
django-autoslug==1.9.9
django-celery-beat==2.5.0
django-celery-email==3.0.0
django-cors-headers==4.1.0
django-countries==7.5.1
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==23.2
django-manifest-loader==1.0.0
django-phonenumber-field==7.1.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.12.2
flower==2.0.0
h11==0.14.0
humanize==4.6.0
idna==3.4
kombu==5.3.1
numpy==1.25.0
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.2
phonenumbers==8.13.15
Pillow==9.5.0
pipenv==2023.6.26
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
sqlparse==0.4.4
tornado==6.3.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
watchdog==3.0.0
wcwidth==0.2.6
webdriver-manager==3.8.6
whitenoise==6.5.0
wsproto==1.2.0
